fixing types

+++ src/AXI_C2C_loop_back_test.cxx
@@ -4,8 +4,8 @@
 
 int SPEED_TEST::AXI_C2C_loop_back_test(string node, uint64_t loops)
 {
-  uint64_t write_mem;
-  uint64_t read_mem;
+  uint32_t write_mem;
+  uint32_t read_mem;
 
   double speed;
